+++ cli/internal/utils/utils.go
@@ -2,6 +2,7 @@ package utils
 
 import (
 	"fmt"
+	"net"
 	"os/exec"
 	"runtime"
 )
@@ -18,4 +19,13 @@ func OpenBrowser(url string) error {
 	default:
 		return fmt.Errorf("unsupported platform")
 	}
+}
+
+func GetCallbackServer() (*net.TCPListener, error) {
+	listener, err := net.Listen("tcp", "localhost:0")
+	if err != nil {
+		return nil, fmt.Errorf("failed to start a callback server: %w", err)
+	}
+
+	return listener.(*net.TCPListener), nil
 }
